refactoring validation v31 fat
