+++ dev/io.openliberty.jakarta.validation.v31_fat/bnd.bnd
@@ -15,7 +15,7 @@ bVersion=1.0
 
 src: \
 	fat/src,\
-	test-applications/app1/src
+	test-applications/val31/src
 
 fat.project: true
