+++ dev/io.openliberty.jakarta.validation.v31_fat/publish/servers/FATServer/server.xml
@@ -19,6 +19,6 @@
     
     <include location="../fatTestPorts.xml"/>
     
-    <application location="app1.war" />
+    <application location="val31.war" />
 
 </server>
